+++ src/main/java/kitchenpos/menus/tobe/domain/MenuProducts.java
@@ -1,0 +1,41 @@
+package kitchenpos.menus.tobe.domain;
+
+import kitchenpos.menus.domain.MenuProduct;
+
+import javax.persistence.*;
+import java.math.BigDecimal;
+import java.util.List;
+
+@Embeddable
+public class MenuProducts {
+    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
+    @JoinColumn(
+            name = "menu_id",
+            nullable = false,
+            columnDefinition = "binary(16)",
+            foreignKey = @ForeignKey(name = "fk_menu_product_to_menu")
+    )
+    private List<MenuProduct> menuProducts;
+
+    public MenuProducts(List<MenuProduct> menuProducts) {
+        this.menuProducts = menuProducts;
+    }
+
+    public MenuProducts() {
+
+    }
+
+    public List<MenuProduct> getMenuProducts() {
+        return menuProducts;
+    }
+
+    public BigDecimal price(){
+        BigDecimal sum = BigDecimal.ZERO;
+        for (final MenuProduct menuProduct : menuProducts) {
+            sum = sum.add(
+                    menuProduct.price()
+            );
+        }
+        return sum;
+    }
+}
